Don't use another MATH constant as a fallback to a MATH constant
For some constants, the OpenType MATH specification suggests to use
the same value as another constant when designing a font. However it
does not make sense to use it as a fallback value when the
MATH.MathConstants is not accessible at all. Instead, use the same
suggested fallback value as the other constant.

https://github.com/mathml-refresh/mathml/issues/69