Rewrite section B.1 Operator Dictionary...
... and the algorithm for determining the properties of an embellished
operator. There should be no behavior change. However section B.1 now
describes two algorithms:

- to determine the category of an operator
- to set the properties of an operator from its category

which are used by the algorithm to set the properties of an embellished
operator. This makes it more convenient for implementations and will
allow to work around the issue with multiple forms tried on operators
that can be both in the Default category for the infix form and in a
different category for a non-infix form.

https://github.com/w3c/mathml-core/issues/104